requirements.txt
asgiref==3.7.2
Django==5.0
django-flat-responsive==2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
uritemplate==4.1.1
